+++ internal/hostenv/osdetect_other.go
@@ -1,0 +1,7 @@
+//go:build !windows
+
+package hostenv
+
+func windowsVersion() string {
+	return ""
+}